Move FormChangeTracker to components/common
