+++ airbyte-webapp/src/pages/ConnectionPage/pages/ConnectionItemPage/ConnectionTransformationTab/DbtCloudTransformationsCard.tsx
@@ -5,7 +5,7 @@ import { Field, Form, Formik, FieldArray, FieldProps, FormikHelpers } from "form
 import { Link } from "react-router-dom";
 import { array, object, number } from "yup";
 
-import { FormChangeTracker } from "components/FormChangeTracker";
+import { FormChangeTracker } from "components/common/FormChangeTracker";
 import { Button } from "components/ui/Button";
 import { Card } from "components/ui/Card";
 import { Input } from "components/ui/Input";
